Include only bank account numbers when search for party
If we search on bank account, this trigger an infinite loop because bank
account search on bank which search back on party
(introduced in changeset 1311ecee55f1).
But indeed when searching for a party, at best it is the bank account number
that is used not the bank or the currency.

issue8991
review288861002